backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
supabase>=2.28.3
python-jose[cryptography]>=3.5.0
httpx>=0.28.0